requirements.txt
sentence-transformers
faiss-cpu
groq
python-dotenv
openai
streamlit
streamlit-carousel
pandas
numpy
torch
torchaudio
transformers
pyaudio
tqdm
tensorflow
tf-keras
python-dotenv
Wave